Create README.md
# Kickstarting with Excel
A project that shows real life scenarios of how data can be analyzed and visualized to make great decision.

## Overview of Project
The Project is about a Louise's play Fever Fund-raising campaign and the project overview is to understand how and when the campaign started, how much fund the campaign will need to succeed, how much each campaign made and if the campaign met their goal.

### Purpose
The purpose of this project is to understand how Louise's play Fever different campaigns faired in relations to their Launch Dates and funding goals. How data can be organized, understood and can be used to penetrate to come up with results and solutions for this fund-raising campaign.

## Analysis and Challenges
Organizing the data and being able the arrange where the data should be, and the interpretations and result expected from the arrangement.
Sorting the data, for the data to be in way that when formulas are applied, desirable result can be obtained.
Visualizing the data, the data set by using charts and graph to understand the trend or pattern at which a data set appears.

### Analysis of Outcomes Based on Launch Date.
A line chart is used so as to help determine trends.
From the chart, for Successful rate, it tried to rise in February, dropped in March, started rising from April and got to its peak in May. The chart shows that, Successful campaigns is highest in the month of May, followed by month of June. Successful campaigns peak was in the May and lowest in December.
For failed campaign, it was practically on the downside; Failed is fairly constant all through the year. The month with the highest failed campaign is October.
For cancelled campaigns, the number of cancelled events, is low the entire month except in October, where there was no cancelled event.


### Analysis of Outcomes Based on Goals
Outcomes based on Goals, Percentage Successful and Percentage Failed seemed to be in opposite direction, when Percentage Failure is going high, the Percentage Failure is going down. Percentage Cancelled does not have interpretation.
When Goal is less than 1000, Percentage successful drops.
At a higher goal of 45000 to 49999, Percentage Successful is at 0% and Percentage Failed at 100%.


### Challenges and Difficulties Encountered
The challenges and difficulties encountered are as follows
1. Understanding the software and where the challenges will go.
2. Challenges with the downloading the file.
3. Understanding the right formula and typing it in correctly.

## Results
A successful project, analyzing and visualizing the data helped to understand the trend of Successful campaigns, Failed Campaigns, and cancelled campaigns. Also showed how much being realized and the total number and rate of participation for the campaigns.


- What are two conclusions you can draw about the Outcomes based on Launch Date?
Conclusions due to Launch date are as follows
1. Successful campaigns were more than failed and cancelled campaigns.
2. Successful Campaigns and Failed Campaigns were at the same point in December.

- What can you conclude about the Outcomes based on Goals?
The Percentage Successful and Percentage Failed trend in opposite direction. 
Percentage Successful reduces as the goal is getting high.

- What are some limitations of this dataset?
I believe the data set does not have recent data. The latest data is in 2017, a lot of things may have changed with date.


- What are some other possible tables and/or graphs that we could create?
Bar Charts, Histogram, Stalked Column, Scattered Charts could have been created.